AbstractFile is created
