+++ src/PGN/File/Syntax.php
@@ -12,23 +12,24 @@ use PGNChess\PGN\Validate;
  * @link https://programarivm.com
  * @license GPL
  */
-class Syntax
+class Syntax extends AbstractFile
 {
+    private $invalid;
+
+    public function __construct()
+    {
+        $invalid = [];
+    }
+
+    /**
+     * Checks if the syntax of a PGN file is valid.
+     *
+     * @return mixed array|bool
+     */
     public function check()
     {
         // TODO ...
 
         return true;
-    }
-
-    private function hasStrTags($tags)
-    {
-        return isset($tags[Tag::EVENT]) &&
-            isset($tags[Tag::SITE]) &&
-            isset($tags[Tag::DATE]) &&
-            isset($tags[Tag::ROUND]) &&
-            isset($tags[Tag::WHITE]) &&
-            isset($tags[Tag::BLACK]) &&
-            isset($tags[Tag::RESULT]);
     }
 }
